+++ tests/Bridge/Symfony/Normalizer/DataCollection.php
@@ -12,4 +12,14 @@ class DataCollection extends AbstractObjectCollection
     {
         parent::__construct(Data::class);
     }
+
+    public function callDoGet(string|int $key): object
+    {
+        return $this->doGet($key);
+    }
+
+    public function add(object $value): static
+    {
+        return $this->doAdd($value);
+    }
 }