+++ tests/Bridge/Symfony/Normalizer/IntCollectionDenormalizerTest.php
@@ -7,7 +7,7 @@ namespace Steevanb\PhpCollection\Tests\Bridge\Symfony\Normalizer;
 use PHPUnit\Framework\TestCase;
 use Steevanb\PhpCollection\{
     Bridge\Symfony\Normalizer\ScalarCollectionDenormalizer,
-    ScalarCollection\IntCollection
+    ScalarCollection\IntegerCollection
 };
 use Symfony\Component\Serializer\Serializer;
 
@@ -16,11 +16,11 @@ final class IntCollectionDenormalizerTest extends TestCase
     public function testDenormalize(): void
     {
         $serializer = new Serializer([new ScalarCollectionDenormalizer()]);
-        /** @var IntCollection $collection */
-        $collection = $serializer->denormalize([1, 2], IntCollection::class);
+        /** @var IntegerCollection $collection */
+        $collection = $serializer->denormalize([1, 2], IntegerCollection::class);
 
         static::assertCount(2, $collection);
-        static::assertSame(1, $collection[0]);
-        static::assertSame(2, $collection[1]);
+        static::assertSame(1, $collection->get(0));
+        static::assertSame(2, $collection->get(1));
     }
 }