+++ tests/Unit/AbstractCollection/AbstractCollectionTest.php
@@ -13,18 +13,18 @@ final class AbstractCollectionTest extends TestCase
         $collection = new Collection([0 => 1, 1 => '2', 2 => null, 'foo' => 'foo']);
 
         static::assertSame([0, 1, 2, 'foo'], array_keys($collection->toArray()));
-        static::assertSame(1, $collection[0]);
-        static::assertSame('2', $collection[1]);
-        static::assertNull($collection[2]);
-        static::assertSame('foo', $collection['foo']);
+        static::assertSame(1, $collection->callDoGet(0));
+        static::assertSame('2', $collection->callDoGet(1));
+        static::assertNull($collection->callDoGet(2));
+        static::assertSame('foo', $collection->callDoGet('foo'));
 
         $collection->resetKeys();
 
         static::assertSame([0, 1, 2, 3], array_keys($collection->toArray()));
-        static::assertSame(1, $collection[0]);
-        static::assertSame('2', $collection[1]);
-        static::assertNull($collection[2]);
-        static::assertSame('foo', $collection[3]);
+        static::assertSame(1, $collection->callDoGet(0));
+        static::assertSame('2', $collection->callDoGet(1));
+        static::assertNull($collection->callDoGet(2));
+        static::assertSame('foo', $collection->callDoGet(3));
     }
 
     public function testToArray(): void