+++ tests/Unit/AbstractCollection/ClearTest.php
@@ -13,17 +13,14 @@ final class ClearTest extends TestCase
         $collection = new Collection();
 
         static::assertCount(0, $collection);
-        static::assertSame(0, $collection->getNextIntKey());
 
         $collection->clear();
 
         static::assertCount(0, $collection);
-        static::assertSame(0, $collection->getNextIntKey());
 
-        $collection[] = 1;
+        $collection->callDoAdd(1);
 
         static::assertCount(1, $collection);
-        static::assertSame(1, $collection->getNextIntKey());
     }
 
     public function testOneItem(): void
@@ -31,17 +28,14 @@ final class ClearTest extends TestCase
         $collection = new Collection([1]);
 
         static::assertCount(1, $collection);
-        static::assertSame(1, $collection->getNextIntKey());
 
         $collection->clear();
 
         static::assertCount(0, $collection);
-        static::assertSame(0, $collection->getNextIntKey());
 
-        $collection[] = 2;
+        $collection->callDoAdd(2);
 
         static::assertCount(1, $collection);
-        static::assertSame(1, $collection->getNextIntKey());
     }
 
     public function testThreeItem(): void
@@ -49,16 +43,13 @@ final class ClearTest extends TestCase
         $collection = new Collection([1, 2, 3]);
 
         static::assertCount(3, $collection);
-        static::assertSame(3, $collection->getNextIntKey());
 
         $collection->clear();
 
         static::assertCount(0, $collection);
-        static::assertSame(0, $collection->getNextIntKey());
 
-        $collection[] = 4;
+        $collection->callDoAdd(4);
 
         static::assertCount(1, $collection);
-        static::assertSame(1, $collection->getNextIntKey());
     }
 }