+++ tests/Unit/ObjectCollection/AbstractObjectCollectionTest.php
@@ -30,10 +30,10 @@ final class AbstractObjectCollectionTest extends TestCase
             ]
         );
 
-        static::assertInstanceOf(TestObject::class, $collection[0]);
-        static::assertSame('foo', $collection[0]->getValue());
-        static::assertInstanceOf(TestObject::class, $collection[1]);
-        static::assertSame('bar', $collection[1]->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(0));
+        static::assertSame('foo', $collection->get(0)->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(1));
+        static::assertSame('bar', $collection->get(1)->getValue());
     }
 
     public function testCanAddValueInvalidInstanceOf(): void
@@ -62,11 +62,11 @@ final class AbstractObjectCollectionTest extends TestCase
         );
 
         static::assertCount(2, $collection);
-        static::assertInstanceOf(TestObject::class, $collection[0]);
-        static::assertSame('foo', $collection[0]->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(0));
+        static::assertSame('foo', $collection->get(0)->getValue());
         // @see https://github.com/steevanb/php-collection/issues/15
-        static::assertInstanceOf(TestObject::class, $collection[2]);
-        static::assertSame('bar', $collection[2]->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(2));
+        static::assertSame('bar', $collection->get(2)->getValue());
     }
 
     public function testComparisonModeObjectHash(): void
@@ -82,11 +82,11 @@ final class AbstractObjectCollectionTest extends TestCase
         );
 
         static::assertCount(3, $collection);
-        static::assertInstanceOf(TestObject::class, $collection[0]);
-        static::assertSame('foo', $collection[0]->getValue());
-        static::assertInstanceOf(TestObject::class, $collection[1]);
-        static::assertSame('foo', $collection[1]->getValue());
-        static::assertInstanceOf(TestObject::class, $collection[2]);
-        static::assertSame('bar', $collection[2]->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(0));
+        static::assertSame('foo', $collection->get(0)->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(1));
+        static::assertSame('foo', $collection->get(1)->getValue());
+        static::assertInstanceOf(TestObject::class, $collection->get(2));
+        static::assertSame('bar', $collection->get(2)->getValue());
     }
 }