+++ tests/Unit/ScalarCollection/StringNullableCollectionTest.php
@@ -19,7 +19,7 @@ final class StringNullableCollectionTest extends TestCase
         $collection = new StringNullableCollection(['4']);
 
         static::assertCount(1, $collection);
-        static::assertSame('4', $collection[0]);
+        static::assertSame('4', $collection->get(0));
     }
 
     public function testInvalidTypeInt(): void
@@ -48,7 +48,7 @@ final class StringNullableCollectionTest extends TestCase
         $collection = new StringNullableCollection([null]);
 
         static::assertCount(1, $collection);
-        static::assertNull($collection[0]);
+        static::assertNull($collection->get(0));
     }
 
     public function testMergeValueAlreadyExistsAdd(): void
@@ -57,10 +57,10 @@ final class StringNullableCollectionTest extends TestCase
             ->merge(new StringNullableCollection(['bar', 'baz']));
 
         static::assertCount(4, $collection);
-        static::assertSame('foo', $collection[0]);
-        static::assertSame('bar', $collection[1]);
-        static::assertSame('bar', $collection[2]);
-        static::assertSame('baz', $collection[3]);
+        static::assertSame('foo', $collection->get(0));
+        static::assertSame('bar', $collection->get(1));
+        static::assertSame('bar', $collection->get(2));
+        static::assertSame('baz', $collection->get(3));
     }
 
     public function testMergeValueAlreadyExistsDoNotAdd(): void
@@ -69,10 +69,10 @@ final class StringNullableCollectionTest extends TestCase
             ->merge(new StringNullableCollection(['bar', 'baz']));
 
         static::assertCount(3, $collection);
-        static::assertSame('foo', $collection[0]);
-        static::assertSame('bar', $collection[1]);
+        static::assertSame('foo', $collection->get(0));
+        static::assertSame('bar', $collection->get(1));
         // @see https://github.com/steevanb/php-collection/issues/15
-        static::assertSame('baz', $collection[3]);
+        static::assertSame('baz', $collection->get(3));
     }
 
     public function testMergeValueAlreadyExistsException(): void